Clean code

+++ src/main/java/com/parkit/parkingsystem/dao/TicketDAO.java
@@ -55,7 +55,7 @@ public class TicketDAO {
                 ticket.setParkingSpot(parkingSpot);
                 ticket.setId(rs.getInt(2));
                 ticket.setVehicleRegNumber(vehicleRegNumber);
-                ticket.setPrice(rs.getInt(3));
+                ticket.setPrice(rs.getDouble(3));
                 ticket.setInTime(rs.getTimestamp(4));
                 ticket.setOutTime(rs.getTimestamp(5));
             }
@@ -71,7 +71,7 @@ public class TicketDAO {
 
     public int getNumberOfTicket(String vehicleRegNumber) {
         Connection con = null;
-       int numberOfTicket = 0;
+        int numberOfTicket = 0;
         try {
             con = dataBaseConfig.getConnection();
             PreparedStatement ps = con.prepareStatement(DBConstants.GET_NB_TICKET);
@@ -108,6 +108,4 @@ public class TicketDAO {
         }
         return false;
     }
-
-
 }
